Fix Stage 5 Optional import + Stage 6 fp16→fp32 dtype cast
Stage 5: add 'from typing import Optional' (missing import)
Stage 6: cast hidden_states to fp32 before LayerNorm in classification head.
  4-bit NF4 model returns fp16 hidden states; head modules initialized in fp32;
  LayerNorm requires consistent dtype — cast once at the entry to the head.

+++ scripts/stage6_distillation/distill.py
@@ -293,9 +293,9 @@ class ClassificationHead(nn.Module):
         self.head  = nn.Linear(128, n_classes)
 
     def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
-        # hidden_states: [B, S, H]
-        # Mean pool over sequence
-        x = hidden_states.mean(dim=1)              # [B, H]
+        # hidden_states: [B, S, H] — may be fp16 from 4-bit model
+        # Cast to float32 for numerical stability in LayerNorm
+        x = hidden_states.float().mean(dim=1)      # [B, H], fp32
         x = self.norm(x)
         x = self.drop(x)
         x = self.act(self.proj(x))
